requirements.txt
streamlit
pandas
numpy
biopython
plotly
